Add Async Make and use posix shell syntax
Introduce an Asyncronous Make command, and a way to kill that Make
process, in order to take advantage of the async things that Vim 8
offers. The commands are Make and MakeK, the functions they use are
Async_Make_Start and Async_Make_Stop respectively.

Use POSIX shell syntax when editing shell scripts that are filetype sh.